+++ tests/test_cli_scene.py
@@ -565,7 +565,10 @@ class TestCliScene(unittest.TestCase):
                 (
                     "Available templates: TEMPLATE.SCENE.LIVE.YOU_ARE_THERE, "
                     "TEMPLATE.SCENE.STEREO.BAND_WIDE_VOCAL_CENTER, "
-                    "TEMPLATE.SCENE.SURROUND.FRONT_STAGE_CLEAR_REAR_FIELD"
+                    "TEMPLATE.SCENE.SURROUND.FRONT_STAGE_CLEAR_REAR_FIELD, "
+                    "TEMPLATE.SEATING.BAND.IN_BAND, "
+                    "TEMPLATE.SEATING.ORCHESTRA.IN_ORCHESTRA, "
+                    "TEMPLATE.SEATING.ORCHESTRA_AUDIENCE"
                 ),
                 first_stderr.getvalue(),
             )